Revert "Architecture: Project restructuring (#1)"
This reverts commit e0be259202da4956b34fedf5923bfaf9433b8f19.
