+++ go.mod
@@ -1,4 +1,4 @@
-module lazyhetzner
+module lazyhetzner.com
 
 go 1.24.4
