--- internal/input_form/input_form.go
@@ -1,14 +1,0 @@
-package input_form
-
-import ( 
-	"github.com/charmbracelet/bubbles/textinput"
-)
-
-
-// Input forms
-type InputForm struct {
-	Inputs    []textinput.Model
-	FocusIdx  int
-	SubmitBtn string
-	CancelBtn string
-}
