--- internal/message/shared_messages.go
@@ -1,16 +1,0 @@
-package message
-
-// Shared common messages for the application
-type ErrorMsg struct {
-	Err error
-}
-
-func (e ErrorMsg) Error() string { return e.Err.Error() }
-
-type StatusMsg string
-
-type ClipboardCopiedMsg string
-
-
-
-
