--- internal/model/state.go
@@ -1,13 +1,0 @@
-package model
-// App states
-type state int
-
-const (
-	StateProjectSelect state = iota
-	stateProjectManage
-	stateTokenInput
-	stateLoading
-	stateResourceView
-	stateContextMenu
-	stateError
-)
